[TASK] Replace classname strings with ::CLASS constants
This replaces the use of class names in strings with using the ::CLASS
constant instead. This has the advantage of immediately being able to
spot errors (the class will be marked as undefined in any proper IDE)
and when refactoring the classes can be identified precisely. Also it
makes the use if namespace imports possible.